requirements.txt
torch
tensorflow
opencv-python
pillow
matplotlib
streamlit
